Move inspec artifact to plugins v2.
