--- lib/bundles/inspec-artifact.rb
@@ -1,7 +1,0 @@
-# encoding: utf-8
-# author: Dave Parfitt
-
-libdir = File.dirname(__FILE__)
-$LOAD_PATH.unshift(libdir) unless $LOAD_PATH.include?(libdir)
-
-require 'inspec-artifact/cli'
