--- lib/bundles/inspec-artifact/README.md
@@ -1,1 +1,0 @@
-# TODO
